requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
sqlalchemy==2.0.23
pytest==7.4.3
pytest-asyncio==0.21.1 